tight_compressor: fix oversampling bug
Fixed bug which affected audio when oversampling was enabled. This bug was introduced in 0.19 when adding a sidechain mode. The channel used for measurement was not participating in oversampling resulting in a small delay between what was being metered and compressed. This is expected to have an effect mostly with short attacks when oversampling.